Make EventTableSchema a subclass of TableSchema, we'll eventually need FeatureTableSchema as its own thing.
